-switch back to pure c++
git-svn-id: http://svn.codergrid.de/svn/sf/bruteripper/trunk@815 7add9c9f-332d-0410-adcc-f7b1da7091d4

+++ src/org/esb/net/SocketInputStream.cpp
@@ -128,7 +128,7 @@ namespace org
         
         /******************************************************************************/
 /*Receive length of buffer*/
-		int available(bool isBlocking){
+		long long int available(bool isBlocking){
 			char tmp[10];
 			_socket->read_some(boost::asio::buffer(tmp),error);
 			if (error == boost::asio::error::eof)
